refactor(#59): работа с таблице в БД pattern вынесена в отдельный модуль
