+++ amplicode.xml
@@ -9,6 +9,7 @@
             </persistence-unit>
             <persistence-unit moduleName="analyzer" name="Default"/>
             <persistence-unit moduleName="core" name="Default"/>
+            <persistence-unit moduleName="pattern-repository" name="Default"/>
         </persistence-units>
     </component>
 </project>
