+++ analyzer/src/main/java/io/github/ijlijapol/bybit/MarketPatternAnalyzer.java
@@ -1,5 +1,7 @@
 package io.github.ijlijapol.bybit;
 
+import io.github.ijlijapol.Pattern;
+import io.github.ijlijapol.PatternRepository;
 import io.github.ijlijapol.bybit.model.Symbol;
 import io.github.ijlijapol.bybit.model.request.LastTime;
 import io.github.ijlijapol.bybit.model.request.RecentMarketDataRequest;
@@ -19,7 +21,7 @@ import java.util.*;
 public class MarketPatternAnalyzer {
 
     private final LoaderMarketData loaderMarketData = MarketDataFactory.getByBitStockMarket();
-    private final MarketPatternRepository patternRepository;
+    private final PatternRepository patternRepository;
 
     public boolean analyzeMarketdata() {
         log.info("Начало анализа рыночных данных за последние пять лет");
@@ -27,7 +29,7 @@ public class MarketPatternAnalyzer {
         final CandlesDTO candlesDTO = getCandlesDTO();
         final List<List<Boolean>> potentialPatterns = getPotentialPatterns(candlesDTO.getCandles());
         final Map<List<Boolean>, Integer> sortedPotentialPatterns = sortPatterns(potentialPatterns);
-        final MarketPattern pattern = getActualPattern(sortedPotentialPatterns);
+        final Pattern pattern = getActualPattern(sortedPotentialPatterns);
 
         patternRepository.save(pattern);
         return true;
@@ -97,7 +99,7 @@ public class MarketPatternAnalyzer {
         return sortedPotentialPatterns;
     }
 
-    private MarketPattern getActualPattern(final Map<List<Boolean>, Integer> sortedPotentialPatterns) {
+    private Pattern getActualPattern(final Map<List<Boolean>, Integer> sortedPotentialPatterns) {
         log.debug("Получение актуального паттерна.");
 
         List<Boolean> actualPattern = new ArrayList<>();
@@ -113,6 +115,6 @@ public class MarketPatternAnalyzer {
         }
 
         log.info("Найден актуальный паттерн: {}", actualPattern);
-        return MarketPattern.builder().candleDirections(actualPattern).build();
+        return Pattern.builder().candleDirections(actualPattern).build();
     }
 }
