--- analyzer/src/main/java/io/github/ijlijapol/bybit/MarketPatternRepository.java
@@ -1,8 +1,0 @@
-package io.github.ijlijapol.bybit;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-@Repository
-public interface MarketPatternRepository extends JpaRepository<MarketPattern, Integer> {
-}
