--- core/src/main/java/io/github/ijlijapol/model/Pattern.java
@@ -1,54 +1,0 @@
-package io.github.ijlijapol.model;
-
-import jakarta.persistence.*;
-import lombok.*;
-import org.hibernate.proxy.HibernateProxy;
-
-import java.util.List;
-import java.util.Objects;
-
-@Builder
-@AllArgsConstructor
-@NoArgsConstructor
-@ToString
-@Getter
-@Setter
-@Entity
-@Table(name = "pattern")
-public class Pattern {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Integer id;
-
-    @ElementCollection(fetch = FetchType.EAGER)
-    @OrderColumn(name = "position")
-    @CollectionTable(
-            name = "pattern_candle_directions",  // явно указываем имя таблицы
-            joinColumns = @JoinColumn(name = "pattern_id")
-    )
-    @Column(name = "candle_directions")
-    private List<Boolean> candleDirections;
-
-    @Override
-    public final boolean equals(Object o) {
-        if (this == o) {
-            return true;
-        }
-        if (o == null) {
-            return false;
-        }
-        Class<?> objectEffectiveClass = o instanceof HibernateProxy proxy ? proxy.getHibernateLazyInitializer().getPersistentClass() : o.getClass();
-        Class<?> thisEffectiveClass = this instanceof HibernateProxy proxy ? proxy.getHibernateLazyInitializer().getPersistentClass() : this.getClass();
-        if (thisEffectiveClass != objectEffectiveClass) {
-            return false;
-        }
-        Pattern pattern = (Pattern) o;
-        return getId() != null && Objects.equals(getId(), pattern.getId());
-    }
-
-    @Override
-    public final int hashCode() {
-        return this instanceof HibernateProxy proxy ? proxy.getHibernateLazyInitializer().getPersistentClass().hashCode() : getClass().hashCode();
-    }
-}
