--- core/src/main/java/io/github/ijlijapol/repostiory/PatternRepository.java
@@ -1,7 +1,0 @@
-package io.github.ijlijapol.repostiory;
-
-import io.github.ijlijapol.model.Pattern;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface PatternRepository extends JpaRepository<Pattern, Integer> {
-}
