+++ docs/design-v1.0.1.md
@@ -96,10 +96,10 @@
 
 #### Список дублирующихся классов
 
-| Класс     | Модуль 1 | Модуль 2      |  Модуль 3  | Различия                                                                                                                                                                                                                                                                                                   | Решение                                                                                                                                                                                                                                                                                                                                                                                                                                     |
-|:----------|:---------|:--------------|:----------:|:-----------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------|:--------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------|
-| `Order`   | `Common` | `Trader`      |   `---`    | Класс `Order` из модуля `Trader`, используется для связи с сущностью в базе данных                                                                                                                                                                                                                         | Класс `Order` из модуля `Common` переименовать в класс `OrderDTO` и оставить там только необходимые поля. Класс `Order` из модуля `Trader` оставить без изменений.                                                                                                                                                                                                                                                                          |
-| `Pattern` | `Trader` | `Test-Trader` | `Analyzer` | Класс `Pattern` имеет различные реализация в разных модулях. Среди некоторыых различий нет, среди других различи в количестве полей. По факту все эти разновидности делятся на сущность БД и `DTO`. Просто из-за особенностей архитекутры приложения пришлось дублировать меняя лишь наименования классов. | Дублирующие классы в модулях `Trader` и `Test-Trader` удалятся при слияние данных двух модулей. Останется только проблема с тем, что данная сущность нужна модулю `Analyzer` и `Core`. На данный момент было принято решение создать новый модуль репозиторный, конкрентно для данной сущности, модули `Core` и `Analyzer` будут подключать репозиторный модуль как зависимость. В дальнейшенм данное решение возможно будет пересмотренно. |
+| Класс                         | Модуль 1 | Модуль 2      |  Модуль 3  | Различия                                                                                                                                                                                                                                                                                                                       | Решение                                                                                                                                                                                                                                                                                                                                                                                                                                     |
+|:------------------------------|:---------|:--------------|:----------:|:-------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------|:--------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------|
+| `Order`                       | `Common` | `Trader`      |   `---`    | Класс `Order` из модуля `Trader`, используется для связи с сущностью в базе данных                                                                                                                                                                                                                                             | Класс `Order` из модуля `Common` переименовать в класс `OrderDTO` и оставить там только необходимые поля. Класс `Order` из модуля `Trader` оставить без изменений.                                                                                                                                                                                                                                                                          |
+| `io.github.ijlijapol.Pattern` | `Trader` | `Test-Trader` | `Analyzer` | Класс `io.github.ijlijapol.Pattern` имеет различные реализация в разных модулях. Среди некоторыых различий нет, среди других различи в количестве полей. По факту все эти разновидности делятся на сущность БД и `DTO`. Просто из-за особенностей архитекутры приложения пришлось дублировать меняя лишь наименования классов. | Дублирующие классы в модулях `Trader` и `Test-Trader` удалятся при слияние данных двух модулей. Останется только проблема с тем, что данная сущность нужна модулю `Analyzer` и `Core`. На данный момент было принято решение создать новый модуль репозиторный, конкрентно для данной сущности, модули `Core` и `Analyzer` будут подключать репозиторный модуль как зависимость. В дальнейшенм данное решение возможно будет пересмотренно. |
 
 #### Правила унификации
