+++ pattern-repository/src/main/java/io/github/ijlijapol/PatternRepository.java
@@ -1,0 +1,6 @@
+package io.github.ijlijapol;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface PatternRepository extends JpaRepository<Pattern, Long> {
+}
